+++ src/main/java/com/github/burgerguy/hudtweaks/util/gui/XAxisNode.java
@@ -11,8 +11,8 @@ public interface XAxisNode {
 	public Set<XAxisNode> getXChildren();
 	public void moveXUnder(XAxisNode newXParent);
 	
-	public int getX(MinecraftClient client);
-	public int getWidth(MinecraftClient client);
+	public double getX(MinecraftClient client);
+	public double getWidth(MinecraftClient client);
 	
 	public void tryUpdateX(UpdateEvent event, MinecraftClient client, boolean parentUpdated, Set<XAxisNode> updatedElementsX);
 }
